Improve JSON class layout (#19)
* Improve JSON classes

* No null for parsed video JSON

* Update todos

* Clearer code

* Ignore .DS_Store files at startup too

* Simplify GetFiles() call

* Clarity

* Tagger minor improvements

* Add creator to comments

* Move app name in comments
