+++ CCVTAC/CCVTAC.Console/PostProcessing/YouTubeJsonExtensionMethods.cs
@@ -5,15 +5,15 @@ public static class YouTubeJsonExtensionMethods
     /// <summary>
     /// Returns a string summarizing video uploader information.
     /// </summary>
-    public static string UploaderSummary(this YouTubeVideoJson.Root data)
+    public static string UploaderSummary(this VideoMetadata videoData)
     {
-        string uploaderLinkOrId = !string.IsNullOrWhiteSpace(data.UploaderUrl)
-            ? data.UploaderUrl
-            : !string.IsNullOrWhiteSpace(data.UploaderId)
-                ? data.UploaderId
+        string uploaderLinkOrId = !string.IsNullOrWhiteSpace(videoData.UploaderUrl)
+            ? videoData.UploaderUrl
+            : !string.IsNullOrWhiteSpace(videoData.UploaderId)
+                ? videoData.UploaderId
                 : string.Empty;
 
-        return data.Uploader +
+        return videoData.Uploader +
                (string.IsNullOrWhiteSpace(uploaderLinkOrId) ? string.Empty : $" ({uploaderLinkOrId})");
     }
 
@@ -21,34 +21,37 @@ public static class YouTubeJsonExtensionMethods
     /// Returns a formatted MM/DD/YYYY version of the upload date (e.g., "08/27/2023") from the
     /// plain YYYYMMDD version (e.g., "20230827") within the parsed JSON file data.
     /// </summary>
-    public static string FormattedUploadDate(this YouTubeVideoJson.Root data) =>
-        $"{data.UploadDate[4..6]}/{data.UploadDate[6..8]}/{data.UploadDate[0..4]}";
+    public static string FormattedUploadDate(this VideoMetadata videoData) =>
+        $"{videoData.UploadDate[4..6]}/{videoData.UploadDate[6..8]}/{videoData.UploadDate[0..4]}";
 
     /// <summary>
     /// Returns a formatted comment using data parsed from the JSON file.
     /// </summary>
-    public static string GenerateComment(
-        this YouTubeVideoJson.Root data,
-        YouTubeCollectionJson.Root? collectionJson)
+    public static string GenerateComment(this VideoMetadata videoData, CollectionMetadata? maybeCollectionData)
     {
         System.Text.StringBuilder sb = new();
-        sb.AppendLine("SOURCE DATA:");
-        sb.AppendLine($"• Downloaded: {DateTime.Now} using CCVTAC");
-        sb.AppendLine($"• Service: {data.ExtractorKey}");
-        sb.AppendLine($"• URL: {data.WebpageUrl}");
-        sb.AppendLine($"• Title: {data.Fulltitle}");
-        sb.AppendLine($"• Uploader: {data.UploaderSummary()}");
-        sb.AppendLine($"• Uploaded: {data.FormattedUploadDate()}");
-        if (collectionJson is not null)
+        sb.AppendLine("CCVTAC SOURCE DATA:");
+        sb.AppendLine($"• Downloaded: {DateTime.Now}");
+        sb.AppendLine($"• Service: {videoData.ExtractorKey}");
+        sb.AppendLine($"• URL: {videoData.WebpageUrl}");
+        sb.AppendLine($"• Title: {videoData.Fulltitle}");
+        sb.AppendLine($"• Uploader: {videoData.UploaderSummary()}");
+        if (videoData.Creator != videoData.Uploader && !string.IsNullOrWhiteSpace(videoData.Creator))
         {
-            sb.AppendLine($"• Playlist name: {collectionJson.Title}");
-            sb.AppendLine($"• Playlist URL: {collectionJson.WebpageUrl}");
-            if (!string.IsNullOrWhiteSpace(collectionJson.Description))
+            sb.AppendLine($"• Creator: {videoData.Creator}");
+        }
+        sb.AppendLine($"• Uploaded: {videoData.FormattedUploadDate()}");
+        if (maybeCollectionData is CollectionMetadata collectionMetadata)
+        {
+            sb.AppendLine($"• Playlist name: {collectionMetadata.Title}");
+            sb.AppendLine($"• Playlist URL: {collectionMetadata.WebpageUrl}");
+            sb.AppendLine($"• Playlist index: {videoData.PlaylistIndex}");
+            if (!string.IsNullOrWhiteSpace(collectionMetadata.Description))
             {
-                sb.AppendLine($"• Playlist description: {collectionJson.Description}");
+                sb.AppendLine($"• Playlist description: {collectionMetadata.Description}");
             }
         }
-        sb.AppendLine($"• Description: {data.Description})");
+        sb.AppendLine($"• Description: {videoData.Description})");
         return sb.ToString();
     }
 }
