+++ CCVTAC/CCVTAC.Console/Settings/UserSettings.cs
@@ -67,7 +67,7 @@ public sealed class UserSettings
     /// If the supplied video uploader is specified in the settings, returns the video's upload year.
     /// Otherwise, returns null.
     /// </summary>
-    public ushort? GetVideoUploadDateIfRegisteredUploader(YouTubeVideoJson.Root videoData)
+    public ushort? GetVideoUploadDateIfRegisteredUploader(VideoMetadata videoData)
     {
         return
             this.UseUploadYearUploaders?.Contains(videoData.Uploader,
