+++ TODO.md
@@ -2,12 +2,7 @@
 
 Features that I might consider as features or improvements in the future. Not all ideas might be viable, or even _good_, but this helps me avoid forgetting them until I decide what to do.
 
-- Ignore, or remove, .DS_Store files in the working directory
-- Avoid using `YouTubeCollectionJson.Root`
-- Do not begin downloading if the working directory is not empty
-- Show current batch number during processing
 - Stop downloading if there are repeated errors from yt-dlp
-- Delete hidden `.DS_Store` files on macOS
 - Logging levels (current/full vs. light)
 - Logging to a file
 - yt-dlp can handle tabs on YouTube channels too, so look into supporting those
